+++ server/src/main/java/au/org/aodn/ogcapi/server/core/model/ogc/wfs/WfsDescribeFeatureTypeResponse.java
@@ -1,4 +1,4 @@
-package au.org.aodn.ogcapi.server.core.model.dto.wfs;
+package au.org.aodn.ogcapi.server.core.model.ogc.wfs;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlElementWrapper;
