+++ server/src/main/java/au/org/aodn/ogcapi/server/core/model/ogc/wfs/WfsInfo.java
@@ -1,4 +1,4 @@
-package au.org.aodn.ogcapi.server.core.model.wfs;
+package au.org.aodn.ogcapi.server.core.model.ogc.wfs;
 
 public record WfsInfo(String wfsUrl, String layerName) {
 }
